128-NEA1 implemented

+++ native/crypto-native/src/tr_havelsan_ueransim_crypto_UIA2.cpp
@@ -1,4 +1,4 @@
-#include "tr_havelsan_ueransim_crypto_UIA2.h"
+#include <jni.h>
 #include "utils.h"
 #include "uea2.h"
 
@@ -7,7 +7,7 @@ extern "C" JNIEXPORT jint JNICALL Java_tr_havelsan_ueransim_crypto_UIA2_computeM
     uint8_t *K = Utils::jbyteArrayToUint8Array(env, key, 1, nullptr);
     uint8_t *M = Utils::jbyteArrayToUint8Array(env, message, 1, nullptr);
 
-    uint32_t mac = UEA2::f9(K, static_cast<uint32_t>(count), static_cast<uint32_t>(fresh), static_cast<uint32_t>(direction), M, bitLength);
+    uint32_t mac = UEA2::f9(K, static_cast<uint32_t>(count), static_cast<uint32_t>(fresh), static_cast<uint32_t>(direction), M, static_cast<uint64_t>(bitLength));
 
     delete[] K;
     delete[] M;
